+++ src/Symfony/Component/DependencyInjection/Tests/Fixtures/php/services9_compiled.php
@@ -203,6 +203,7 @@ class ProjectServiceContainer extends Container
 
         $instance->setBar($this->get('foo'));
         $instance->setBar(NULL);
+        $instance->setBar(($this->get("foo")->foo() . $this->getParameter("foo")));
 
         return $instance;
     }